Changed the structure of the repository, integrated Catch2 and Cmake
Decided to make changes to the structure of the repository the better
reflect its purpose and potential use.

This repository is intended for an educational purpose for data structures and algorithms.
Each such implementation is self-contained in its subdirectory of the lib directory and
can be used as a library with its public API header located at the include directory.

Each such library is now tested using the Catch2 C/C++ test framework with test files
located at the test directory.

All compilation of the code(into static libraries) and the test files is now handled by Cmake.